feat(core/executionWindows): add endpoint to get next time for an execution window
