+++ orca-web/src/main/groovy/com/netflix/spinnaker/orca/controllers/TimeWindowController.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2018 Netflix, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License")
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.netflix.spinnaker.orca.controllers;
+
+import com.netflix.spinnaker.orca.pipeline.RestrictExecutionDuringTimeWindow.ExecutionWindowConfig;
+import com.netflix.spinnaker.orca.pipeline.window.TimeWindowCalculator;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.time.Instant;
+
+@RestController
+public class TimeWindowController {
+
+  @Autowired
+  private TimeWindowCalculator timeWindowCalculator;
+
+  @RequestMapping(value = "/timeWindow/nextAvailable", method = RequestMethod.POST)
+  public long getNextAvailableTimeWindow(@RequestBody ExecutionWindowConfig windowConfig) {
+    return timeWindowCalculator.getTimeInWindow(windowConfig, Instant.now()).toEpochMilli();
+  }
+
+}
